backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.7.0
bleach==6.0.0
Bottleneck==1.3.7
certifi==2023.5.7
cffi==1.15.1
click==8.1.5
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.0
fastjsonschema==2.17.1
fqdn==1.5.1
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
llvmlite==0.40.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.1
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.1
numexpr==2.8.4
numpy==1.24.4
odfpy==1.4.1
openpyxl==3.1.2
orjson==3.9.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.0.3
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.3.0
Pygments==2.15.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
pyxlsb==1.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.29.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.8.10
scipy==1.11.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
stack-data==0.6.2
starlette==0.27.0
terminado==0.17.1
tinycss2==1.2.1
tokenize-rt==5.1.0
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
uri-template==1.3.0
uvicorn==0.23.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
widgetsnbextension==4.0.8
xarray==2023.6.0
xlrd==2.0.1
XlsxWriter==3.1.2
